sprint2/step1 pr fixes
