+++ src/components/app/app.jsx
@@ -80,7 +80,6 @@ function App() {
             )}
         </main>
       </div>
-      <div id="react-modals" />
     </>
   );
 }
